Add DuckDB integration tests and helper functions for query management
- Introduced duckdbRefreshHelpers.js with functions for entering, saving, and asserting queries.
- Created duckdbRefreshPersistence.cy.js to test DuckDB persistence scenarios, including handling empty results and refreshing queries.
- Added duckdbRefreshReconciliation.cy.js for testing DuckDB reconciliation with BigQuery, including source renaming and public viewer scenarios.
- Updated updateDataset.cy.js to extend timeout for query readiness checks.
- Enhanced local DuckDB tests to improve editor interaction and SQL assertion methods.
